Checking timescale being passed into cook frame - synchronized objects seem to be running too fast This commit has a print screen for the time in milliseconds * 10 being passed into the touch engine instance

+++ Source/TouchEngine/Public/TouchEngineComponent.h
@@ -54,7 +54,7 @@ protected:
 	FDelegateHandle paramsLoadedDelHandle, loadFailedDelHandle;
 	FDelegateHandle beginFrameDelHandle, endFrameDelHandle;
 
-	int64 cookTime = 0, lastCookTime = 0;
+	float cookTime = 0, lastCookTime = 0;
 
 	// Called when the game starts
 	virtual void BeginPlay() override;
